requirements.txt
streamlit==1.31.0
supabase==2.3.4
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
